refactor(game): 주요 클래스를 components 폴더로 분리한다
- Car.js
- Game.js
- Input.js
- Output.js
